+++ POD_Blum-Micali_Generator/POD_Blum-Micali_Generator/Model/SiG.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Numerics;
 using System.Text;
 using System.Threading.Tasks;
 
@@ -9,7 +10,7 @@ namespace POD_Blum_Micali_Generator.Model
     //Generator 0 1
     class SiG
     {
-        public int Si(UInt64 xn, UInt64 p)
+        public int Si(BigInteger xn, BigInteger p)
         {
             if (xn > ((p - 1) / 2))
             {
